backend/requirement.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
email-validator==2.3.0
fastapi==0.135.1
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib[bcrypt]==1.7.4
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-jose==3.5.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
python-multipart

